+++ src/test/java/nextstep/subway/acceptance/AuthAcceptanceTest.java
@@ -43,24 +43,4 @@ class AuthAcceptanceTest extends AcceptanceTest {
 
         회원_정보_조회됨(response, EMAIL, AGE);
     }
-
-    public static ExtractableResponse<Response> 폼_로그인_후_내_회원_정보_조회_요청(String email, String password) {
-        return RestAssured
-            .given().log().all()
-            .auth().form(email, password, new FormAuthConfig("/login/form", USERNAME_FIELD, PASSWORD_FIELD))
-            .accept(MediaType.APPLICATION_JSON_VALUE)
-            .when().get("/members/me")
-            .then().log().all()
-            .statusCode(HttpStatus.OK.value()).extract();
-    }
-
-    public static ExtractableResponse<Response> 베어러_인증으로_내_회원_정보_조회_요청(String accessToken) {
-        return RestAssured.given().log().all()
-                          .auth().oauth2(accessToken)
-                          .accept(MediaType.APPLICATION_JSON_VALUE)
-                          .when().get("/members/me")
-                          .then().log().all()
-                          .statusCode(HttpStatus.OK.value())
-                          .extract();
-    }
 }
